Show loading icon when changing to next or previous image
I tried to replace the image with a loading icon instead, but that
caused the image to blink for every image switch. Having a loading
icon for the button which is pressed seem to work better.

It is important to have a loading icon as the images might be quite
big.