Change from storing next_image_number to store current_image_number
The reason for this is that a page that shows the current image for an
album is to be created.

It would have been possible to find the current image number by
subtracting 1 from the next image number, but there was currently no
guarantee that this image existed. Such a guarantee is now
implemented.