Change to using homepage in package.json
The problem with the previous approach was that manually entered urls
would not work when the react app was served from flask. It was
possible to go to /albums/album1, but refreshing the page caused the
screen to turn white. Somehow the problem seems so be solved when
using homepage instead.