Create slideshow page at album/<album_name>/slideshow
The slideshow page is now set to gather data every 3 minutes. The
reason for this long delay is that every data-gathering causes a
rerender of the slideshow component. This in turn causes the interval
of the slideshow component to be reset, which creates an extra
transition in the slideshow. (This will look weird to a viewer)

If it exists a way to not get the extra transition, the data gathering
interval could be set lower than 3 minutes.

BTW, the slideshow works as follows:

1. The image in front is set to the same as the image in the back
2. After 1 second, the image in the back is switched out
3. After 1 additional second (two seconds from when the front image
was set) the front image fades out
4. repeat (after some time)